+++ oommfc/tests/test_mindriver.py
@@ -25,6 +25,9 @@ class TestMinDriver(TestDriver):
         assert "Oxs_MinDriver" in script
         assert "Oxs_FileVectorField" in script
 
+        if os.path.exists('tds'):
+            shutil.rmtree('tds')
+
     @pytest.mark.oommf
     def test_drive(self):
         md = oc.MinDriver()
@@ -45,4 +48,5 @@ class TestMinDriver(TestDriver):
 
         assert len(odt_files) == 1
 
-        shutil.rmtree("tds")
+        if os.path.exists('tds'):
+            shutil.rmtree('tds')
